+++ app/.deps/esper-run
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+python manage.py run $1
