+++ app/query/datasets/babycam/models.py
@@ -1,5 +1,6 @@
 from django.db import models
 import query.base_models as base
+import math
 
 class Video(base.Video):
     session_id = models.IntegerField()
@@ -12,6 +13,9 @@ class Video(base.Video):
     context_country = base.CharField()
     context_state = base.CharField()
 
+    def get_stride(self):
+        return int(math.ceil(self.fps)/2)
+
 class Frame(base.Frame):
     pass
 
@@ -19,4 +23,7 @@ class Labeler(base.Labeler):
     pass
 
 class Face(base.Concept):
+    pass
+
+class Pose(base.Concept, base.Pose):
     pass
