+++ app/query/datasets/tvnews/models.py
@@ -27,26 +27,5 @@ class Face(base.Concept):
     gender = models.ForeignKey(Gender, on_delete=models.CASCADE, null=True, blank=True)
     identity = models.ForeignKey(Identity, on_delete=models.CASCADE, null=True, blank=True)
 
-class Pose(base.Concept):
-    keypoints = models.BinaryField()
-
-    def _format_keypoints(self):
-        kp = np.frombuffer(self.keypoints, dtype=np.float32)
-        return kp.reshape((kp.shape[0]/3, 3))
-
-    POSE_KEYPOINTS = 18
-    FACE_KEYPOINTS = 70
-    HAND_KEYPOINTS = 21
-
-    def pose_keypoints(self):
-        kp = self._format_keypoints()
-        return kp[:self.POSE_KEYPOINTS, :]
-
-    def face_keypoints(self):
-        kp = self._format_keypoints()
-        return kp[self.POSE_KEYPOINTS:(self.POSE_KEYPOINTS+self.FACE_KEYPOINTS), :]
-
-    def hand_keypoints(self):
-        kp = self._format_keypoints()
-        base = kp[self.POSE_KEYPOINTS+self.FACE_KEYPOINTS:, :]
-        return [base[:self.HAND_KEYPOINTS, :], base[self.HAND_KEYPOINTS:, :]]
+class Pose(base.Concept, base.Pose):
+    pass
